requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-tavily
yfinance
rich
pydantic
python-dotenv